+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft CMS 3.x
 
+## Unreleased
+
+- Fixed a bug where `Craft.BaseElementIndexView::this.canSelectElement()` wasn’t getting applied for lazy-loaded elements.
+
 ## 3.8.16 - 2023-07-18
 
 - The “Access the control panel” user permission now includes a warning that the permission grants view-only access to user data and most content.